requirements.txt
pandas
flask 
flask-apscheduler
feedparser
python-dotenv
openai
gptcache
beautifulsoup4
SQLAlchemy
psycopg2-binary
playwright
elevenlabs